A Review on Domain Adaption and Generative Adversarial Networks(GANs)
arXiv: 2510.12075
Authors: Aashish Dhawan, Divyanshu Mudgal
Published: 2025-10-14

The major challenge in today's computer vision scenario is the availability of good quality labeled data. In a field of study like image classification, where data is of utmost importance, we need to find more reliable methods which can overcome the scarcity of data to produce results comparable to previous benchmark results. In most cases, obtaining labeled data is very difficult because of the high cost of human labor and in some cases impossible. The purpose of this paper is to discuss Domain Adaptation and various methods to implement it. The main idea is to use a model trained on a particular dataset to predict on data from a different domain of the same kind, for example - a model trained on paintings of airplanes predicting on real images of airplanes

Links:
- Abstract: https://arxiv.org/abs/2510.12075
- PDF: https://arxiv.org/pdf/2510.12075

Categories: cs.CV, cs.AI