Knowledge and Common Knowledge of Strategies
arXiv: 2510.19298
Authors: Borja Sierra Miranda, Thomas Studer
Published: 2025-10-22

Most existing work on strategic reasoning simply adopts either an informed or an uninformed semantics. We propose a model where knowledge of strategies can be specified on a fine-grained level. In particular, it is possible to distinguish first-order, higher-order, and common knowledge of strategies. We illustrate the effect of higher-order knowledge of strategies by studying the game Hanabi. Further, we show that common knowledge of strategies is necessary to solve the consensus problem. Finally, we study the decidability of the model checking problem.

Links:
- Abstract: https://arxiv.org/abs/2510.19298
- PDF: https://arxiv.org/pdf/2510.19298

Categories: cs.LO, cs.AI